Move all the widgets to a dedicated folder
