+++ phpcs.xml.dist
@@ -191,5 +191,5 @@
 	<!-- Specific to Polylang -->
 	<exclude-pattern>bin/*.php</exclude-pattern>
 	<exclude-pattern>src/install/plugin-updater.php</exclude-pattern>
-	<exclude-pattern>src/widget-calendar.php</exclude-pattern>
+	<exclude-pattern>src/modules/Widgets/widget-calendar.php</exclude-pattern>
 </ruleset>
