+++ src/base.php
@@ -4,8 +4,8 @@
  */
 
 use WP_Syntex\Polylang\REST\Request;
-use WP_Syntex\Polylang\Language_Switcher\Widget;
 use WP_Syntex\Polylang\Capabilities\Capabilities;
+use WP_Syntex\Polylang\Widgets\Languages as Languages_Widget;
 
 /**
  * Base class for both admin and frontend
@@ -118,7 +118,7 @@ abstract class PLL_Base {
 	 */
 	public function widgets_init() {
 		register_widget( 'PLL_Widget_Languages' );
-		register_widget( Widget::class );
+		register_widget( Languages_Widget::class );
 
 		// Overwrites the calendar widget to filter posts by language
 		if ( ! defined( 'PLL_WIDGET_CALENDAR' ) || PLL_WIDGET_CALENDAR ) {
